refactor(replacecr): move replace-cr reader to a small package
I've found a use for this elsewhere, so a package it becomes!
