+++ dsio/csv.go
@@ -1,7 +1,6 @@
 package dsio
 
 import (
-	"bufio"
 	"encoding/csv"
 	"encoding/json"
 	"fmt"
@@ -9,6 +8,7 @@ import (
 	"strconv"
 
 	"github.com/qri-io/dataset"
+	"github.com/qri-io/dataset/dsio/replacecr"
 	"github.com/qri-io/dataset/vals"
 )
 
@@ -25,7 +25,7 @@ func NewCSVReader(st *dataset.Structure, r io.Reader) *CSVReader {
 	// TODO - handle error
 	_, types, _ := terribleHackToGetHeaderRowAndTypes(st)
 
-	csvr := csv.NewReader(ReplaceSoloCarriageReturns(r))
+	csvr := csv.NewReader(replacecr.Reader(r))
 
 	if opts, ok := st.FormatConfig.(*dataset.CSVOptions); ok {
 		csvr.LazyQuotes = opts.LazyQuotes
@@ -273,52 +273,4 @@ func encode(vs []interface{}) ([]string, error) {
 func (w *CSVWriter) Close() error {
 	w.w.Flush()
 	return nil
-}
-
-// ReplaceSoloCarriageReturns wraps an io.Reader, on every call of Read. it looks for
-// for instances of lonely \r replacing them with \r\n before returning to the end consumer
-// lots of files in the wild will come without "proper" line breaks, which irritates go's
-// standard csv package. This'll fix by wrapping the reader passed to csv.NewReader:
-// 		rdr, err := csv.NewReader(ReplaceSoloCarriageReturns(r))
-//
-func ReplaceSoloCarriageReturns(data io.Reader) io.Reader {
-	return crlfReplaceReader{
-		rdr: bufio.NewReader(data),
-	}
-}
-
-// crlfReplaceReader wraps a reader
-type crlfReplaceReader struct {
-	rdr *bufio.Reader
-}
-
-// Read implements io.Reader for crlfReplaceReader
-func (c crlfReplaceReader) Read(p []byte) (n int, err error) {
-	if len(p) == 0 {
-		return
-	}
-
-	for {
-		if n == len(p) {
-			return
-		}
-
-		p[n], err = c.rdr.ReadByte()
-		if err != nil {
-			log.Debug(err.Error())
-			return
-		}
-
-		// any time we encounter \r & still have space, check to see if \n follows
-		// ff next char is not \n, add it in manually
-		if p[n] == '\r' && n < len(p) {
-			if pk, err := c.rdr.Peek(1); (err == nil && pk[0] != '\n') || (err != nil && err.Error() == "EOF") {
-				n++
-				p[n] = '\n'
-			}
-		}
-
-		n++
-	}
-	return
 }
