+++ dsio/csv_test.go
@@ -240,24 +240,6 @@ a	12	23	"[""foo"",""bar""]"`
 		t.Errorf("output mismatch. %s != %s", buf.String(), expect)
 	}
 }
-
-func TestReplaceSoloCarriageReturns(t *testing.T) {
-	input := []byte("foo\r\rbar\r\nbaz\r\r")
-	expect := []byte("foo\r\n\r\nbar\r\nbaz\r\n\r\n")
-
-	got := make([]byte, 19)
-	n, err := ReplaceSoloCarriageReturns(bytes.NewReader(input)).Read(got)
-	if err != nil && err.Error() != "EOF" {
-		t.Errorf("unexpected error: %s", err.Error())
-	}
-	if n != 19 {
-		t.Errorf("length error. expected: %d, got: %d", 19, n)
-	}
-	if !bytes.Equal(expect, got) {
-		t.Errorf("byte mismatch. expected:\n%v\ngot:\n%v", expect, got)
-	}
-}
-
 func BenchmarkCSVWriterArrays(b *testing.B) {
 	const NumWrites = 1000
 	st := &dataset.Structure{Format: dataset.CSVDataFormat, Schema: dataset.BaseSchemaObject}
